test(datetime): migrate disable-dates
